jwt auth system added

+++ index.js
@@ -4,6 +4,7 @@ const cors = require("cors");
 const { MongoClient, ServerApiVersion, ObjectId } = require("mongodb");
 const jwt = require("jsonwebtoken");
 require("dotenv").config();
+const stripe = require("stripe")(process.env.STRIPE_SECRET_API_KEY);
 const app = express();
 const port = process.env.PORT || 5000;
 
@@ -17,6 +18,21 @@ const client = new MongoClient(uri, {
   serverApi: ServerApiVersion.v1,
 });
 
+function verifyJWT(req, res, next) {
+  const auth = req.headers.authorization;
+  if (!auth) {
+    res.status(403).send("forbidden access");
+  }
+  const token = auth.split(" ")[1];
+  jwt.verify(token, process.env.JWT_ACCESS_TOKEN, function (error, decoded) {
+    if (error) {
+      res.status(401).send("unauthorized access");
+    }
+    req.decoded = decoded;
+    next();
+  });
+}
+
 async function run() {
   try {
     const productsCollection = client
@@ -29,9 +45,25 @@ async function run() {
     const soldProductsCollection = client
       .db("recycle-laptop")
       .collection("soldProducts");
-      const wishlistCollection= client.db("recycle-laptop").collection("wishlistProducts");
-      const reportedProductsCollection= client.db("recycle-laptop").collection("reportedProducts");
-
+    const wishlistCollection = client
+      .db("recycle-laptop")
+      .collection("wishlistProducts");
+    const reportedProductsCollection = client
+      .db("recycle-laptop")
+      .collection("reportedProducts");
+    const bookedProductCollection = client
+      .db("recycle-laptop")
+      .collection("bookedProducts");
+///verify admin role
+    const verifyAdmin = async (req, res, next) => {
+      const email = req.decoded.email;
+      const query = { user_email: email };
+      const adminEmail = await usersCollection.findOne(query);
+      if (adminEmail?.user_role !== "admin") {
+        return res.status(403).send("forbidden access");
+      }
+      next();
+    };
     //store new product info
     app.post("/product", async (req, res) => {
       const product = req.body;
@@ -42,13 +74,18 @@ async function run() {
     app.get("/advertised/products", async (req, res) => {
       const query = {
         isAdvertised: true,
+        saleStatus: "unsold",
       };
       const products = await productsCollection.find(query).toArray();
       res.send(products);
     });
     //get seller product
-    app.get("/products", async (req, res) => {
+    app.get("/products",verifyJWT, async (req, res) => {
       const email = req.query.email;
+      const verifyEmail = req.decoded.email;
+      if (verifyEmail !== email) {
+        return res.status(401).send("unauthorized access");
+      }
       let query;
       if (email) {
         query = {
@@ -59,9 +96,13 @@ async function run() {
       res.send(products);
     });
     //delete a product
-    app.delete("/product/delete/:id", async (req, res) => {
+    app.delete("/product/delete/:id",verifyJWT, async (req, res) => {
+      const email = req.query.email;
+      const verifyEmail = req.decoded.email;
+      if (verifyEmail !== email) {
+        return res.status(401).send("unauthorized access");
+      }
       const id = req.params.id;
-
       const query = { _id: ObjectId(id) };
       const result = await productsCollection.deleteOne(query);
       res.send(result);
@@ -84,18 +125,17 @@ async function run() {
       res.send(result);
     });
     //update report product status
-    app.post('/report-product', async(req, res)=>{
+    app.post("/report-product", async (req, res) => {
       const data = req.body;
       const result = await reportedProductsCollection.insertOne(data);
       res.send(result);
-    })
+    });
     //add product wishlist
-    app.post('/wishlist-product', async(req, res)=>{
+    app.post("/wishlist-product", async (req, res) => {
       const data = req.body;
       const result = await wishlistCollection.insertOne(data);
       res.send(result);
-    
-    })
+    });
     //store sold product
     app.post("/sold-product", async (req, res) => {
       const data = req.body;
@@ -103,13 +143,13 @@ async function run() {
       res.send(result);
     });
     //update sold product status
-    app.put("/sold-product/:id", async (req, res) => {
+    app.put("/booked-product/:id", async (req, res) => {
       const id = req.params.id;
       const filter = { _id: ObjectId(id) };
       const options = { upsert: true };
       const updateDoc = {
         $set: {
-          saleStatus: "sold",
+          saleStatus: "booked",
         },
       };
       const result = await productsCollection.updateOne(
@@ -129,7 +169,7 @@ async function run() {
     //get a single categories
     app.get("/single_category/:id", async (req, res) => {
       const id = parseInt(req.params.id);
-      const query = { categoryId: id };
+      const query = { categoryId: id, saleStatus: "unsold" };
       const cursor = await productsCollection.find(query).toArray();
       res.send(cursor);
     });
@@ -140,8 +180,12 @@ async function run() {
       res.send(result);
     });
     //get a user info
-    app.get("/user", async (req, res) => {
+    app.get("/user",verifyJWT, async (req, res) => {
       const email = req.query.email;
+      const verifyEmail = req.decoded.email;
+      if (verifyEmail !== email) {
+        return res.status(401).send("unauthorized access");
+      }
       const query = {
         user_email: email,
       };
@@ -150,24 +194,188 @@ async function run() {
     });
 
     //get all buyer
-    app.get('/all-buyer', async(req, res)=>{
-        const query={user_role:'buyer'}
-        const result=await usersCollection.find(query).toArray();
-        res.send(result)
-    })
+    app.get("/all-buyer", async (req, res) => {
+      const query = { user_role: "buyer" };
+      const result = await usersCollection.find(query).toArray();
+      res.send(result);
+    });
+    //delete a buyer
+    app.delete("/all-buyer/:id",verifyJWT, verifyAdmin, async (req, res) => {
+      const id = req.params.id;
+      const query = { _id: ObjectId(id) };
+      const result = await usersCollection.deleteOne(query);
+      res.send(result);
+    });
     //get all seller
-    app.get('/all-seller', async(req, res)=>{
-        const query={user_role:'seller'}
-        const result=await usersCollection.find(query).toArray();
-        res.send(result)
-    })
-
+    app.get("/all-seller", async (req, res) => {
+      const query = { user_role: "seller" };
+      const result = await usersCollection.find(query).toArray();
+      res.send(result);
+    });
+    app.get("/jwt", async (req, res) => {
+      const email = req.query.email;
+      const query = { user_email: email };
+      const user = usersCollection.findOne(query);
+      if (user) {
+        const token = jwt.sign({ email }, process.env.JWT_ACCESS_TOKEN, {
+          expiresIn: "1d",
+        });
+        return res.send({ accessToken: token });
+      }
+      res.status(401).send("unauthorized access");
+    });
     //get all reported products
-    app.get('/reported-products', async(req, res)=>{
-      const query={};
-      const result=await reportedProductsCollection.find(query).toArray();
-      res.send(result)
-    })
+    app.get("/reported-products", async (req, res) => {
+      const query = {};
+      const result = await reportedProductsCollection.find(query).toArray();
+      res.send(result);
+    });
+    //get all wishlist product
+    app.get("/my-wishlist",verifyJWT, async (req, res) => {
+      const email = req.query.email;
+      const verifyEmail = req.decoded.email;
+      if (verifyEmail !== email) {
+        return res.status(401).send("unauthorized access");
+      }
+      const query = { user_email: email };
+      const wishlist = await wishlistCollection.find(query).toArray();
+      res.send(wishlist);
+    });
+    //get all products
+    app.get("/all-products", async (req, res) => {
+      const query = { saleStatus: "unsold" };
+
+      const products = await productsCollection.find(query).toArray();
+      res.send(products);
+    });
+
+    //payment intent
+
+    app.post("/create-payment-intent", async (req, res) => {
+      const price = req.body.price;
+
+      // Create a PaymentIntent with the order amount and currency
+      const paymentIntent = await stripe.paymentIntents.create({
+        amount: price,
+        currency: "usd",
+        payment_method_types: ["card"],
+      });
+
+      res.send({
+        clientSecret: paymentIntent.client_secret,
+      });
+    });
+
+    //store booked product
+    app.post("/my-orders", async (req, res) => {
+      const data = req.body;
+      const result = await bookedProductCollection.insertOne(data);
+      res.send(result);
+    });
+
+    //update wishlist sold status
+    app.put("/wishlistsProducts/:id", async (req, res) => {
+      const id = req.params.id;
+      const filter = { _id: ObjectId(id) };
+      const options = { upsert: true };
+      const updateDoc = {
+        $set: {
+          saleStatus: "sold",
+        },
+      };
+      const result = await wishlistCollection.updateOne(
+        filter,
+        updateDoc,
+        options
+      );
+      res.send(result);
+    });
+
+    ///get my orders
+    app.get("/my-orders",verifyJWT, async (req, res) => {
+      const email = req.query.email;
+      const verifyEmail = req.decoded.email;
+      if (verifyEmail !== email) {
+        return res.status(401).send("unauthorized access");
+      }
+      const query = { customerEmail: email };
+      const result = await bookedProductCollection.find(query).toArray();
+      res.send(result);
+    });
+
+    //remove item from booked
+    app.delete("/remove-from-cart/:id", verifyJWT, async (req, res) => {
+      const email = req.query.email;
+      const verifyEmail = req.decoded.email;
+      if (verifyEmail !== email) {
+        return res.status(401).send("unauthorized access");
+      }
+      const id = req.params.id;
+      const filter = { _id: ObjectId(id) };
+      const result = await bookedProductCollection.deleteOne(filter);
+      res.send(result);
+    });
+    //remove from wishlist
+    app.delete("/remove-from-wishlist/:id", verifyJWT, async (req, res) => {
+      const email = req.query.email;
+      const verifyEmail = req.decoded.email;
+      if (verifyEmail !== email) {
+        return res.status(401).send("unauthorized access");
+      }
+      const id = req.params.id;
+      const filter = { _id: ObjectId(id) };
+      const result = await wishlistCollection.deleteOne(filter);
+      res.send(result);
+    });
+    //get product for payment
+    app.get("/payment/:id", async (req, res) => {
+      const id = req.params.id;
+      const query = { _id: ObjectId(id) };
+      const result = await productsCollection.findOne(query);
+      res.send(result);
+    });
+    //sold product status changed
+    app.put("/soldProducts/:id", async (req, res) => {
+      const id = req.params.id;
+      const filter = { _id: ObjectId(id) };
+      const options = { upsert: true };
+      const updateDoc = {
+        $set: {
+          saleStatus: "sold",
+        },
+      };
+      const result = await productsCollection.updateOne(
+        filter,
+        updateDoc,
+        options
+      );
+      res.send(result);
+    });
+
+    //update booked product collection sold status
+    app.put("/bookedProducts/:id", async (req, res) => {
+      const id = req.params.id;
+      const filter = { _id: ObjectId(id) };
+      const options = { upsert: true };
+      const updateDoc = {
+        $set: {
+          saleStatus: "sold",
+        },
+      };
+      const result = await bookedProductCollection.updateOne(
+        filter,
+        updateDoc,
+        options
+      );
+      res.send(result);
+    });
+    ///delete from report collection
+    app.delete("/delete-report/:id", async (req, res) => {
+      const id = req.params.id;
+      const query = { _id: ObjectId(id) };
+      const result = await reportedProductsCollection.deleteOne(query);
+      res.send(result);
+    });
   } catch (error) {
     console.log(error);
   }
